Modify multiple files to a specified format

Converts a group files located in the same folder to a new format. It iterates through the provided vector of files and creates a copy of each file in the specified format (or extension). The user can set the location of the new files, which is shared across all files.

Usage

convert_files(
  folder,
  files,
  new_extension = "parquet",
  new_folder = NULL,
  verbose = TRUE
)

Arguments

folder

The path to the folder containing the original files.

files

A character vector with the file paths to be copy. The paths must satisfy the following properties:

The path must start from the folder path.

Have a file extensions supported by read_fun().

new_extension

A string with the format that must have the copies of the original files. Must be a extension supported by write_fun().

new_folder

The path to the folder where the converted files will be saved (if the folder does not exist will be created). If not provided, the new folder will be created inside the original folder with the same name.

verbose

Logical (the default is TRUE) indicating whether to display progress messages.

Value

None.

Details

This function is robust to uppercase and dots in the new extension. If the new_folder argument is not provided, the new folder will be created inside the original folder with the same name. If needed, sub folders will be created inside the new folder to mirror the structure of the original folder.

Examples

if (FALSE) {
# Convert files in the current directory to Parquet format
convert_files(folder = ".", files = list.files(), new_extension = "parquet")
}