Refine the "raw partial dictionary"

This function refine the raw partial dictionary created by create_partial_dictionary(). It reads the raw dictionary file, sorts the column names by frequency (across all files) and alphabetically, and creates some columns with descriptive statistics for each variable. The sorted (refined) dictionary is saved into a xlsx file.

Usage

sort_partial_dictionary(old_dict_path, new_dict_path = NULL, overwrite = F)

Arguments

old_dict_path

Path to the raw dictionary file.

new_dict_path

Path to save the sorted (refined) dictionary If NULL, the original file will be old_dict_path (the default is NULL).

overwrite

Logical indicating whether to overwrite the existing dictionary file if new_dict_path already exists. Its default value is FALSE to avoid undesired changes.

Value

None. The function saves the sorted dictionary file to the specified location. The dictionary has the following columns:

uniname: Suggested unifying name for each variable. It groups across files identical case robust variable names.

uniclass: Empty column that is intended to be filled manually by the user with the unifying class for each variable. Theoretically, this could be filled with any value, however, based on the template in vignette('vignettes/process_data_with_partial_dict.Rmd') is recommended to use 'integer', 'numeric', 'character', 'date' and 'logical'.

coverage: The percentage of files that have a match with the uniname.

class_mode: The class mode per uniname computed with get_mode().

unique_classes: All the classes per uniname.

See also

Examples

if (FALSE) {
# Sort a partial dictionary file
sort_partial_dictionary(old_dict_path = "original_dictionary.xlsx",
                          overwrite = TRUE)
}